@@ -6,38 +6,45 @@
   "scripts": {
     "build": "webpack --mode=production",
     "watch": "webpack --mode=development --watch",
-    "lint": "eslint './src/**/*.js'"
+    "lint": "npx eslint \"./src/scripts/**/*.js\""
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/codefluegel/h5p-editor-virtual-3D-tour.git"
+  },
+  "type": "module",
+  "homepage": "https://github.com/codefluegel/h5p-editor-virtual-3D-tour#readme",
   "devDependencies": {
-    "@babel/cli": "7.26.4",
-    "@babel/core": "^7.26.7",
-    "@babel/eslint-parser": "^7.26.5",
-    "@babel/preset-env": "^7.26.7",
+    "@babel/cli": "7.27.0",
+    "@babel/core": "^7.26.10",
+    "@babel/eslint-parser": "^7.27.0",
+    "@babel/preset-env": "^7.26.9",
     "@babel/preset-react": "^7.26.3",
-    "@types/react": "19.0.0",
-    "@types/react-dom": "19.0.0",
+    "@types/react": "19.1.0",
+    "@types/react-dom": "19.1.1",
     "babel-loader": "^9.2.1",
     "css-loader": "^7.1.2",
-    "eslint": "^9.23.0",
-    "eslint-plugin-jsdoc": "^50.6.3",
-    "eslint-plugin-react": "^7.37.4",
+    "eslint": "^9.24.0",
+    "eslint-plugin-jsdoc": "^50.6.9",
+    "eslint-config-snordian-h5p": "github:otacke/eslint-config-snordian-h5p",
+    "eslint-plugin-react": "^7.37.5",
     "mini-css-extract-plugin": "^2.9.2",
-    "sass": "^1.84.0",
-    "sass-loader": "^16.0.4",
-    "webpack": "^5.97.1",
+    "sass": "^1.86.3",
+    "sass-loader": "^16.0.5",
+    "webpack": "^5.99.0",
     "webpack-cli": "^6.0.1"
   },
   "dependencies": {
-    "@babel/plugin-transform-runtime": "^7.25.9",
+    "@babel/plugin-transform-runtime": "^7.26.10",
     "globals": "^16.0.0",
     "he": "^1.2.0",
     "prop-types": "^15.8.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-toastify": "^11.0.3"
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-toastify": "^11.0.5"
   },
   "keywords": [],
-  "author": "CodeFluegel GmBh",
+  "author": "CodeFluegel",
   "license": "MIT",
   "browserslist": "> 0.5%, last 2 versions, Firefox ESR, not dead"
 }
